refactor: move vnodes to framework folder
